requirements.txt
wheel>=0.23.0
PyICU>=1.8
pycld2>=0.3
six>=1.7.3
futures>=2.1.6
morfessor>=2.0.2a1
numpy>=1.6.1
polyglot
opus-fast-mosestokenizer
mecab-ko
mecab-python3
unidic-lite
opustools
